+++ .storybook/main.js
@@ -1,6 +1,3 @@
 module.exports = {
-  stories: ['../packages/**/stories/**/*.stories.{js,mdx}'],
-  esDevServer: {
-    // custom es-dev-server options
-  },
+  stories: ['../packages/**/stories/**/*.stories.js'],
 };