+++ packages/image-card/stories/index.stories.js
@@ -1,4 +1,3 @@
-import { html } from 'lit-html';
 import '../cmmd-image-card.js';
 
 export default {
@@ -6,24 +5,24 @@ export default {
 };
 
 export const example = () =>
-  html`
+  `
     <cmmd-image-card
       height="360px"
       width="640px"
       src="https://picsum.photos/640/360"
-      title="Lovely image"
+      imageCardTitle="Lovely image"
       alt="This is a lovely image"
     >
     </cmmd-image-card>
   `;
 
 export const button = () =>
-  html`
+  `
     <cmmd-image-card
       height="360px"
       width="640px"
       src="https://picsum.photos/640/360"
-      title="Lovely image"
+      imageCardTitle="Lovely image"
       alt="This is a lovely image"
       .buttonRole="${true}"
     >