+++ packages/styles/src/sizes.js
@@ -1,3 +1,3 @@
-import { css } from 'lit-element';
+import { css } from 'lit';
 
 export const borderRadiusSize = css`0.25rem`;