+++ packages/button/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 2.0.0 (2023-07-21)
+
+
+### Bug Fixes
+
+* **button:** index.js added export ([66e68a6](https://github.com/carlosmmdiaz/cmmd-web/commit/66e68a6aa8f357db681ffd4410f76e2b8195d470))
+
+
+### Features
+
+* 🎸 all components updated to last version of Lit ([980b4e2](https://github.com/carlosmmdiaz/cmmd-web/commit/980b4e230a078e061732e13d0db6db4d73eab480))
+* **button:** component finished ([3311a69](https://github.com/carlosmmdiaz/cmmd-web/commit/3311a69661f38a1bcbe4d7b286046383d80ab7a2))
+* **button:** create button component ([f1b5d1a](https://github.com/carlosmmdiaz/cmmd-web/commit/f1b5d1a1eca665e9f0f5cf1ef6b000860a04cf3a))
+* **button:** use borderRadiusSize ([bb022ae](https://github.com/carlosmmdiaz/cmmd-web/commit/bb022aeb50f059da374ecb50e2a2e3d06f619f22))
+* **button:** use CMMD fontFamily ([7a7859e](https://github.com/carlosmmdiaz/cmmd-web/commit/7a7859e0842b6ba2bb48ff6d063b73f9a7dcb82f))
+
+
+### BREAKING CHANGES
+
+* 🧨 YES
+
+
+
+
+
 # 1.1.0 (2020-04-26)