+++ packages/card/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 2.0.0 (2023-07-21)
+
+
+### Features
+
+* 🎸 all components updated to last version of Lit ([980b4e2](https://github.com/carlosmmdiaz/cmmd-web/commit/980b4e230a078e061732e13d0db6db4d73eab480))
+* **card:** create card component ([348767d](https://github.com/carlosmmdiaz/cmmd-web/commit/348767d5f92c79b719ec5bf0c4caa54a928b97ee))
+* **card:** use borderRadiusSize ([c47bafb](https://github.com/carlosmmdiaz/cmmd-web/commit/c47bafb1a2023736a5287f84d99cb53e7e85cfba))
+
+
+### BREAKING CHANGES
+
+* 🧨 YES
+
+
+
+
+
 # 1.1.0 (2020-04-26)