+++ packages/input/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 2.0.0 (2023-07-21)
+
+
+### Features
+
+* 🎸 all components updated to last version of Lit ([980b4e2](https://github.com/carlosmmdiaz/cmmd-web/commit/980b4e230a078e061732e13d0db6db4d73eab480))
+* **input:** add styles ([d7360b9](https://github.com/carlosmmdiaz/cmmd-web/commit/d7360b9f12f089371ebcafb5c7147d34de21665e))
+* **input:** basic component created ([16ec05e](https://github.com/carlosmmdiaz/cmmd-web/commit/16ec05ebf780c6ffff33492f44e19de988449799))
+* **input:** use borderRadiusSize ([cde4aef](https://github.com/carlosmmdiaz/cmmd-web/commit/cde4aef9261ba28df81c011cdfbe33d38f922a5c))
+
+
+### BREAKING CHANGES
+
+* 🧨 YES
+
+
+
+
+
 # 1.1.0 (2020-04-26)