+++ packages/styles/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 2.0.0 (2023-07-21)
+
+
+### Bug Fixes
+
+* **dependencies:** remove the unnecessary ones ([e641bcf](https://github.com/carlosmmdiaz/cmmd-web/commit/e641bcf323cbd0efe5b85720da3397a93a733b70))
+
+
+### Features
+
+* 🎸 all components updated to last version of Lit ([980b4e2](https://github.com/carlosmmdiaz/cmmd-web/commit/980b4e230a078e061732e13d0db6db4d73eab480))
+* **styles:** add CMMD fonts ([c9b9430](https://github.com/carlosmmdiaz/cmmd-web/commit/c9b94300d46fdf41ed6105b33f1bd2fdb98e85b1))
+* **styles:** add plus and cross icons ([f40e2f9](https://github.com/carlosmmdiaz/cmmd-web/commit/f40e2f91a4f12e96d58987ae65cd6e7f377fdce4))
+* **styles:** add sizes file, add borderRadiusSize ([fae86f8](https://github.com/carlosmmdiaz/cmmd-web/commit/fae86f87bdbf34b2f22a158b41f34b99ab0e08c3))
+* **styles:** add transparentBlack rgba ([a375f76](https://github.com/carlosmmdiaz/cmmd-web/commit/a375f769ee67c4cc1ab242389b2fc71cf7ab5a8f))
+* **styles:** initial commit ([0559e10](https://github.com/carlosmmdiaz/cmmd-web/commit/0559e1024d05065f15fdd09c8edfb585c9dd3977))
+* **styles:** new color darkgrey added ([675ae70](https://github.com/carlosmmdiaz/cmmd-web/commit/675ae703384df5bfeb3f54ecfbde0fccfa2625a1))
+* **styles:** new colors and icon ([0cec6bb](https://github.com/carlosmmdiaz/cmmd-web/commit/0cec6bb7b5b0924531798ee58b8a34b437a2b21a))
+
+
+### BREAKING CHANGES
+
+* 🧨 YES
+
+
+
+
+
 # 1.1.0 (2020-04-26)